requirements.txt
pytorch_lightning
jsonargparse[signatures]
huggingface_hub
transformers
pytorch_optimizer
einops
matplotlib
fairseq
librosa
torchcrepe
pesq
tqdm